refactor: imageRepository 헥사고날 아키텍처에 맞게 변경 (#107)
